requirements.txt
# Requirements for the Streamlit application 
streamlit>=1.28.0
requests>=2.31.0
python-dotenv>=1.0.0